rename utils.test.ts to fixtures.test.ts
