+++ packages/core/src/command/command.fixtures.test.ts
