+++ packages/core/src/eventStore/eventStore.unit.test.ts
@@ -15,7 +15,7 @@ import {
   putSnapshotMock,
   counterCreatedEventMock,
   counterIncrementedEventMock,
-} from './eventStore.util.test';
+} from './eventStore.fixtures.test';
 
 describe('event store', () => {
   beforeEach(() => {
