+++ packages/json-schema-command/src/command.fixtures.test.ts
