+++ packages/json-schema-command/src/command.unit.test.ts
@@ -8,7 +8,7 @@ import {
   inputSchema,
   outputSchema,
   requiredEventStores,
-} from './command.util.test';
+} from './command.fixtures.test';
 
 getEventsMock.mockResolvedValue({ events: counterEventsMocks });
